QuantumDots.jl Documentation

Documentation for QuantumDots.

This package provides some tools for working with quantum systems, especially interacting systems of fermions. It is not registered but can be installed by

using Pkg; Pkg.add(url="https://github.com/cvsvensson/QuantumDots.jl")

Introduction

Let's analyze a small fermionic system. We first define a basis

using QuantumDots
N = 2 # number of fermions
spatial_labels = 1:N 
internal_labels = (:↑,:↓)
c = FermionBasis(spatial_labels, internal_labels)

Indexing into the basis like returns sparse representations of the fermionic operators, so that one can write down Hamiltonians in a natural way

H_hopping = c[1,:↑]'c[2,:↑] + c[1,:↓]'c[2,:↓] + hc 
H_coulomb = sum(c[n,:↑]'c[n,:↑]c[n,:↓]'c[n,:↓] for n in spatial_labels)
H = H_hopping + H_coulomb
#= 16×16 SparseArrays.SparseMatrixCSC{Int64, Int64} with 23 stored entries:
⎡⠠⠂⠀⠀⠀⠀⠀⠀⎤
⎢⠀⠀⠰⢂⠑⢄⠀⠀⎥
⎢⠀⠀⠑⢄⠠⢆⠀⠀⎥
⎣⠀⠀⠀⠀⠀⠀⠰⢆⎦ =#

One can also work in the single particle basis FermionBdGBasis if the system is noninteracting. Quadratic functions of the fermionic operators produce the single particle BdG Hamiltonian.

c2 = FermionBdGBasis(spatial_labels, internal_labels)
Hfree = c2[1,:↑]'c2[2,:↑] + c2[1,:↓]'c2[2,:↓] + hc
vals, vecs = diagonalize(BdGMatrix(Hfree)) 

Using diagonalize on a matrix of type BdGMatrix enforces particle-hole symmetry for the eigenvectors.

More info

For a more in depth introduction see pmm_notebook.

QubitBasis and time evolution is demonstrated in qubit_dephasing.

Simulation of Majorana braiding with noisy gates is demonstrated in majorana_braiding.

Most functionalities of the package are demonstrated in the tests.

QuantumDots.AbelianFockSymmetry

QuantumDots.BdGFermion

QuantumDots.BdGMatrix

QuantumDots.DiagonalizedHamiltonian

QuantumDots.FermionBasis

QuantumDots.KhatriRaoVectorizer

QuantumDots.KronVectorizer

QuantumDots.LindbladCache

QuantumDots.LindbladDissipator

QuantumDots.LindbladSystem

QuantumDots.LindbladSystem

QuantumDots.PauliDissipator

QuantumDots.PauliSystem

QuantumDots.QuasiParticle

QuantumDots.QuasiParticle

Base.:*

Base.:*

QuantumDots.bdg_to_skew

QuantumDots.blockdiagonal

QuantumDots.dissipator

QuantumDots.embedding_unitary

QuantumDots.eval_in_basis

QuantumDots.fermion_sparse_matrix

QuantumDots.fermionic_embedding

QuantumDots.fockstates

QuantumDots.focksymmetry

QuantumDots.jwstring

QuantumDots.majorana_coefficients

QuantumDots.many_body_density_matrix

QuantumDots.many_body_fermion

QuantumDots.many_body_hamiltonian

QuantumDots.normalized_steady_state_rhs

QuantumDots.numberoperator

QuantumDots.one_particle_density_matrix

QuantumDots.one_particle_density_matrix

QuantumDots.one_particle_density_matrix

QuantumDots.one_particle_density_matrix

QuantumDots.ordered_prod_of_embeddings

QuantumDots.parityoperator

QuantumDots.partial_trace

QuantumDots.partial_trace!

QuantumDots.partial_transpose

QuantumDots.quasiparticle_adjoint

QuantumDots.ratetransform

QuantumDots.ratetransform

QuantumDots.removefermion

QuantumDots.rep

QuantumDots.rotate_majorana_coefficients

QuantumDots.skew_to_bdg

QuantumDots.skew_to_bdg

QuantumDots.superoperator

QuantumDots.wedge

QuantumDots.wedge

QuantumDots.@fermions

QuantumDots.@majoranas

QuantumDots.AbelianFockSymmetry — Type

struct AbelianFockSymmetry{IF,FI,QN,QNfunc} <: AbstractSymmetry

AbelianFockSymmetry represents a symmetry that is diagonal in fock space, i.e. particle number conservation, parity, spin consvervation.

Fields

indtofockdict::IF: A dictionary mapping indices to Fock states.

focktoinddict::FI: A dictionary mapping Fock states to indices.

qntoblocksizes::Dictionary{QN,Int}: A dictionary mapping quantum numbers to block sizes.

qntofockstates::Dictionary{QN,Vector{Int}}: A dictionary mapping quantum numbers to Fock states.

qntoinds::Dictionary{QN,Vector{Int}}: A dictionary mapping quantum numbers to indices.

conserved_quantity::QNfunc: A function that computes the conserved quantity from a fock number.

QuantumDots.BdGFermion — Type

struct BdGFermion{S,B,T} <: AbstractBdGFermion

The BdGFermion struct represents a basis fermion for BdG matrices.

Fields

id::S: The identifier of the fermion.

basis::B: The fermion basis.

amp::T: The amplitude of the fermion (default: true, i.e. 1).

hole::Bool: Indicates whether the fermion is a hole (default: true).

QuantumDots.BdGMatrix — Type

struct BdGMatrix <: AbstractMatrix

BdGMatrix represents a Bogoliubov-de Gennes (BdG) matrix, which is a block matrix used to describe non-interacting superconducting systems. It consists of four blocks: H, Δ, -conj(Δ), and -conj(H), where H is a Hermitian matrix and Δ is an antisymmetric matrix.

Fields

H: The Hermitian block of the BdG matrix.

Δ: The antisymmetric block of the BdG matrix.

QuantumDots.DiagonalizedHamiltonian — Type

struct DiagonalizedHamiltonian{Vals,Vecs,H} <: AbstractDiagonalHamiltonian

A struct representing a diagonalized Hamiltonian.

Fields

values: The eigenvalues of the Hamiltonian.

vectors: The eigenvectors of the Hamiltonian.

original: The original Hamiltonian.

QuantumDots.FermionBasis — Type

struct FermionBasis{M,D,Sym,L} <: AbstractManyBodyBasis

Fermion basis for representing many-body fermions.

Fields

dict::OrderedDict: A dictionary that maps fermion labels to a representation of the fermion.

symmetry::Sym: The symmetry of the basis.

jw::JordanWignerOrdering{L}: The Jordan-Wigner ordering of the basis.

QuantumDots.KhatriRaoVectorizer — Type

struct KhatriRaoVectorizer{T} <: AbstractVectorizer

A struct representing a Khatri-Rao vectorizer. This vectorizer is used for BlockDiagonal density matrices matrices, where the superoperators respect the block structure.

Fields

sizes::Vector{Int}: Vector of sizes.

idvec::Vector{T}: Vector of identifiers.

cumsum::Vector{Int}: Vector of cumulative sums.

cumsumsquared::Vector{Int}: Vector of squared cumulative sums.

inds::Vector{UnitRange{Int}}: Vector of index ranges.

vectorinds::Vector{UnitRange{Int}}: Vector of vector index ranges.

QuantumDots.KronVectorizer — Type

struct KronVectorizer{T} <: AbstractVectorizer

A struct representing a KronVectorizer, the standard vectorizer where superoperators are formed from kronecker products of operators.

Fields

size::Int: The size of the KronVectorizer.

idvec::Vector{T}: A vector representing the vectorized identity matrix. Saved here because it is useful when normalization is needed for computations in e.g. LindbladSystem.

QuantumDots.LindbladCache — Type

struct LindbladCache{KC, MC, SC, OC}

A cache structure used in the Lindblad equation solver.

Fields

kroncache::KC: Cache for Kronecker products.

mulcache::MC: Cache for matrix multiplications.

superopcache::SC: Cache for superoperators.

opcache::OC: Cache for operators.

QuantumDots.LindbladDissipator — Type

struct LindbladDissipator{S,T,L,H,V,C} <: AbstractDissipator

A struct representing a Lindblad dissipator.

Fields

superop::S: The superoperator representing the dissipator.

rate::T: The rate of the dissipator.

lead::L: The lead associated with the dissipator.

ham::H: The Hamiltonian associated with the dissipator.

vectorizer::V: The vectorizer used for vectorization.

cache::C: The cache used for storing intermediate results.

QuantumDots.LindbladSystem — Type

LindbladSystem(hamiltonian, leads, vectorizer=default_vectorizer(hamiltonian); rates=Dict(k => 1 for (k, v) in pairs(leads)), usecache=false)

Constructs a Lindblad system for simulating open quantum systems.

Arguments

hamiltonian: The Hamiltonian of the system.

leads: An list of operators representing the leads.

vectorizer: Determines how to vectorize the lindblad equation. Defaults to default_vectorizer(hamiltonian).

rates: An array of rates for each lead. Defaults to an array of ones with the same length as leads.

usecache: A boolean indicating whether to use a cache. Defaults to false.

QuantumDots.LindbladSystem — Type

struct LindbladSystem{T,U,DS,V,H,C} <: AbstractOpenSystem

A struct representing a Lindblad open quantum system.

Fields

total::T: The total lindblad matrix operator.

unitary::U: The unitary part of the system.

dissipators::DS: The dissipators of the system.

vectorizer::V: The vectorizer used for the system.

hamiltonian::H: The Hamiltonian of the system.

cache::C: The cache used for the system.

QuantumDots.PauliDissipator — Method

PauliDissipator(ham, lead; change_lead_basis=true)

Constructs the Pauli dissipator for a given Hamiltonian and lead.

Arguments

ham: The Hamiltonian.

lead: The leads.

change_lead_basis: A boolean indicating whether to change the lead basis to the energy basis. Default is true.

QuantumDots.PauliSystem — Method

PauliSystem(ham, leads)

Constructs a PauliSystem object from a Hamiltonian and a set of leads.

QuantumDots.QuasiParticle — Type

struct QuasiParticle{T,M,L} <: AbstractBdGFermion

The QuasiParticle struct represents a quasi-particle in the context of a BdG (Bogoliubov-de Gennes) fermion system. It is a linear combination of basis BdG fermions, and is defined by a set of weights and a basis.

QuantumDots.QuasiParticle — Method

QuasiParticle(f::BdGFermion)

Constructs a QuasiParticle object from a BdGFermion object.

Base.:* — Method

*(f1::BdGFermion, f2::BdGFermion; symmetrize=true)

Multiply two BdGFermion objects f1 and f2. By default, it symmetrizes the result, returning a BdG matrix in the convention used here.

Base.:* — Method

*(f1::QuasiParticle, f2::QuasiParticle; kwargs...)

Return the BdG matrix of the product of quasiparticles f1 and f2.

QuantumDots.bdg_to_skew — Method

bdg_to_skew(B::BdGMatrix; check=true)

Convert a BdGMatrix to a skew-Hermitian matrix. If check is true, it checks that the result is skew-Hermitian.

QuantumDots.blockdiagonal — Method

blockdiagonal(m::AbstractMatrix, basis::AbstractManyBodyBasis)

Construct a BlockDiagonal version of m using the symmetry of basis. No checking is done to ensure this is a faithful representation.

QuantumDots.dissipator — Method

dissipator(L, rate, kv::KronVectorizer)

Constructs the dissipator associated to the jump operator L.

QuantumDots.embedding_unitary — Method

embedding_unitary(partition, fockstates, jw)

Compute the unitary matrix that maps between the tensor embedding and the fermionic embedding in the physical subspace. 
# Arguments
- `partition`: A partition of the labels in `jw` into disjoint sets.
- `fockstates`: The fock states in the basis
- `jw`: The Jordan-Wigner ordering.

QuantumDots.eval_in_basis — Method

eval_in_basis(a, f::AbstractBasis)

Evaluate an expression with fermions in a basis f.

Examples

@fermions a
f = FermionBasis(1:2)
QuantumDots.eval_in_basis(a[1]'*a[2] + hc, f)

QuantumDots.fermion_sparse_matrix — Method

fermion_sparse_matrix(fermion_number, totalsize, sym)

Constructs a sparse matrix of size totalsize representing a fermionic operator at bit position fermion_number in a many-body fermionic system with symmetry sym.

QuantumDots.fermionic_embedding — Method

fermionic_embedding(m, b, bnew)

Compute the fermionic embedding of a matrix m in the basis b into the basis bnew.

QuantumDots.fockstates — Method

fockstates(M, n)

Generate a list of Fock states with n occupied fermions in a system with M different fermions.

QuantumDots.focksymmetry — Method

focksymmetry(fockstates, qn)

Constructs an AbelianFockSymmetry object that represents the symmetry of a many-body fermionic system.

Arguments

fockstates: The fockstates to iterate over

qn: A function that takes an integer representing a fock state and returns corresponding quantum number.

QuantumDots.jwstring — Method

jwstring(site, focknbr)

Parity of the number of fermions to the right of site.

QuantumDots.majorana_coefficients — Function

majorana_coefficients(f::QuasiParticle, labels=collect(keys((basis(f))))

Compute the Majorana coefficients for a given QuasiParticle object f. Returns two dictionaries, for the two types of Majorana operators.

QuantumDots.many_body_density_matrix — Function

many_body_density_matrix(G, c=FermionBasis(1:div(size(G, 1), 2), qn=parity); alg=SkewEigenAlg())

Compute the many-body density matrix for a given correlator G. The traceless version of G should be a BdGMatrix.

See also one_particle_density_matrix, many_body_hamiltonian.

QuantumDots.many_body_fermion — Method

many_body_fermion(f::BdGFermion, basis::FermionBasis)

Return the representation of f in the many-body fermion basis basis.

QuantumDots.many_body_hamiltonian — Function

many_body_hamiltonian(H::AbstractMatrix, Δ::AbstractMatrix, c::FermionBasis=FermionBasis(1:size(H, 1), qn=parity))

Construct the many-body Hamiltonian for a given BdG Hamiltonian consisting of hoppings H and pairings Δ.

QuantumDots.normalized_steady_state_rhs — Method

normalized_steady_state_rhs(A)

For a linear operator A, whose last row represents the normalization condition, return the rhs of the equation A * x = b where x is the normalized steady-state and b the vector with all zeros except for the last element which is 1.

QuantumDots.numberoperator — Method

numberoperator(basis::FermionBasis)

Return the number operator of basis.

QuantumDots.one_particle_density_matrix — Method

one_particle_density_matrix(χs::AbstractVector{<:QuasiParticle})

Return the oneparticledensity_matrix for the state with all χs occupied as its ground state.

QuantumDots.one_particle_density_matrix — Method

one_particle_density_matrix(U::AbstractMatrix{T}) where {T}

Compute the one-particle density matrix for the vacuum of a BdG system diagonalized by U.

QuantumDots.one_particle_density_matrix — Method

one_particle_density_matrix(χ::QuasiParticle{T})

Return the one_particle_density_matrix for the state with χ occupied as its ground state.

QuantumDots.one_particle_density_matrix — Method

one_particle_density_matrix(ρ::AbstractMatrix, b::FermionBasis, labels=keys(b))

Compute the one-particle density matrix for a given density matrix ρ in the many body fermion basis b.

QuantumDots.ordered_prod_of_embeddings — Method

ordered_prod_of_embeddings(ms, bs, b)

Compute the ordered product of the fermionic embeddings of the matrices ms in the bases bs into the basis b.

QuantumDots.parityoperator — Method

parityoperator(basis::AbstractBasis)

Return the parity operator of basis.

QuantumDots.partial_trace! — Function

partial_trace!(mout, m::AbstractMatrix, b::AbstractManyBodyBasis, bout::AbstractManyBodyBasis, phase_factors)

Compute the fermionic partial trace of a matrix m in basis b, leaving only the subsystems specified by labels. The result is stored in mout, and bout determines the ordering of the basis states.

QuantumDots.partial_trace — Method

partial_trace(m::AbstractMatrix,  bfull::AbstractBasis, bsub::AbstractBasis)

Compute the partial trace of a matrix m, leaving the subsystem defined by the basis bsub.

QuantumDots.partial_transpose — Function

partial_transpose(m::AbstractMatrix, b::AbstractManyBodyBasis, labels)

Compute the fermionic partial transpose of a matrix m in subsystem denoted by labels.

QuantumDots.quasiparticle_adjoint — Method

quasiparticle_adjoint(v::AbstractVector)

Compute the adjoint of a quasiparticle represented by the weights in v. The adjoint is computed by swapping the hole and particle parts of the vector and taking the complex conjugate of each element.

QuantumDots.ratetransform — Method

ratetransform(op,  energies::AbstractVector, T, μ)

Transform op in the energy basis with a Fermi-Dirac distribution at temperature T and chemical potential μ.

Arguments

op: The operator to be transformed.

energies::AbstractVector: The energies.

T: The temperature.

μ: The chemical potential.

QuantumDots.ratetransform — Method

ratetransform(op, diagham::DiagonalizedHamiltonian, T, μ)

Transform op with a Fermi-Dirac distribution at temperature T and chemical potential μ.

Arguments

op: The operator to be transformed.

diagham::DiagonalizedHamiltonian: The diagonalized Hamiltonian.

T: The temperature.

μ: The chemical potential.

QuantumDots.removefermion — Method

removefermion(digitposition, statefocknbr)

Return (newfocknbr, fermionstatistics) where newfocknbr is the state obtained by removing a fermion at digitposition from statefocknbr and fermionstatistics is the phase from the Jordan-Wigner string, or 0 if the operation is not allowed.

QuantumDots.rep — Method

rep(f::BdGFermion)

Constructs a sparse vector representation of a BdGFermion object.

QuantumDots.rotate_majorana_coefficients — Method

rotate_majorana_coefficients(w, z, phase)

Rotate the Majorana coefficients by `phase`

QuantumDots.skew_to_bdg — Method

skew_to_bdg(A::AbstractMatrix)

Convert a skew-symmetric matrix A to a BdGMatrix.

QuantumDots.skew_to_bdg — Method

skew_to_bdg(v::AbstractVector)

Use the same transformation that transforms a skew-symmetric matrix to a BdGMatrix to transform a vector v.

QuantumDots.superoperator — Method

superoperator(lead_op, diagham, T, μ, rate, vectorizer)

Construct the superoperator associated with the operator lead_op. Transforms the operator to the energy basis and includes fermi-Dirac statistics.

Arguments

lead_op: The operator representing the lead coupling.

diagham: The diagonal Hamiltonian.

T: The temperature.

μ: The chemical potential.

rate: The rate of the dissipative process.

vectorizer: The vectorizer struct.

QuantumDots.wedge — Function

wedge(ms::AbstractVector, bs::AbstractVector{<:FermionBasis}, b::FermionBasis=wedge(bs))

Compute the wedge product of matrices or vectors in ms with respect to the fermion bases bs, respectively. Return a matrix in the fermion basis b, which defaults to the wedge product of bs.

QuantumDots.wedge — Method

wedge(bs)

Compute the wedge product of a list of FermionBasis objects. The symmetry of the resulting basis is computed by promote_symmetry.

QuantumDots.@fermions — Macro

@fermions a b ...

Create one or more fermion species with the given names. Indexing into fermions species gives a concrete fermion. Fermions in one @fermions block anticommute with each other, and commute with fermions in other @fermions blocks.

Examples:

@fermions a b creates two species of fermions that anticommute:

a[1]' * a[1] + a[1] * a[1]' == 1

a[1]' * b[1] + b[1] * a[1]' == 0

@fermions a; @fermions b creates two species of fermions that commute with each other:

a[1]' * a[1] + a[1] * a[1]' == 1

a[1] * b[1] - b[1] * a[1] == 0

See also @majoranas, QuantumDots.eval_in_basis.

QuantumDots.@majoranas — Macro

@majoranas a b ...

Create one or more Majorana species with the given names. Indexing into Majorana species gives a concrete Majorana. Majoranas in one @majoranas block anticommute with each other, and commute with Majoranas in other @majoranas blocks.

Examples:

@majoranas a b creates two species of Majoranas that anticommute:

a[1] * a[1] + a[1] * a[1] == 1

a[1] * b[1] + b[1] * a[1] == 0

@majoranas a; @majoranas b creates two species of Majoranas that commute with each other:

a[1] * a[1] + a[1] * a[1] == 1

a[1] * b[1] - b[1] * a[1] == 0

See also @fermions, QuantumDots.eval_in_basis.